% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata_and_counts.R
\name{calculate_pseudobulk}
\alias{calculate_pseudobulk}
\title{Generate pseudobulk counts and quantile_normalised counts}
\usage{
calculate_pseudobulk(
  sce_data,
  atlas_name,
  import_date,
  cache_dir = get_default_cache_dir()
)
}
\arguments{
\item{sce_data}{A SingleCellExperiment object from RDS, the metadata slot of which
must contain \code{cell_} and \code{dataset_id}}

\item{atlas_name}{A character string specifying the name of the atlas to import.}

\item{import_date}{A character vector that specifies the date of the import.
The date should be in the international format 'YYYY-MM-DD' for clarity and consistency.}

\item{cache_dir}{Optional character vector of length 1. A file path on
your local system to a directory (not a file) that will be used to store pseudobulk counts}
}
\value{
Pseudobulk counts in \code{Anndata} format stored in the cache directory
}
\description{
Generate pseudobulk counts and quantile_normalised counts
}
