% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{check_gene_overlap}
\alias{check_gene_overlap}
\title{Checks whether genes in a list of SummarizedExperiment objects overlap}
\usage{
check_gene_overlap(obj_list)
}
\arguments{
\item{obj_list}{A list of SummarizedExperiment objects}
}
\value{
A character vector of genes intersection across objects
}
\description{
Checks whether genes in a list of SummarizedExperiment objects overlap
}
\keyword{internal}
