% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_old_metadata}
\alias{clear_old_metadata}
\title{Clear the outdated metadata in the default cache directory.}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
clear_old_metadata(metadata)
}
\arguments{
\item{updated_data}{A character vector of outdated metadata name}
}
\value{
\code{NULL}, invisibly
}
\description{
Clear the outdated metadata in the default cache directory.
}
\keyword{internal}
