% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_app.R
\name{create_interface_app}
\alias{create_interface_app}
\title{Create a Shiny app that allows users to generate filtering & retreival code for cellNexus}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
create_interface_app(ui_choices, return_as_list = FALSE)
}
\arguments{
\item{ui_choices}{A list of pre-computed unique values for each filterable column in the metadata.}

\item{return_as_list}{If TRUE, returns a list with 'ui' and 'server' components instead of a Shiny app object.}
}
\value{
A Shiny app that allows users to filter cellNexus metadata and generate code for retrieval
in the selected format.
}
\description{
Create a Shiny app that allows users to generate filtering & retreival code for cellNexus
}
\examples{
get_default_cache_dir()
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Create the interface app with metadata
metadata <- get_metadata(cloud_metadata = SAMPLE_DATABASE_URL)
app <- create_interface_app(metadata)
# Run the app
shiny::runApp(app)
\dontshow{\}) # examplesIf}
}
\author{
Jared Andrews
}
