% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_counts}
\alias{delete_counts}
\title{Deletes specific counts and metadata from cache}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
delete_counts(
  data,
  assay = c("original", "cpm"),
  cache_directory = get_default_cache_dir()
)
}
\value{
\code{NULL}, invisibly
}
\description{
Deletes specific counts and metadata from cache
}
\keyword{internal}
