% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{duckdb_read_parquet}
\alias{duckdb_read_parquet}
\title{Returns a tibble from a parquet file path via DuckDB
Since dbplyr 2.4.0, raw file paths aren't handled very well
See: \url{https://github.com/duckdb/duckdb-r/issues/38}
Hence the need for this method}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
duckdb_read_parquet(conn, path, filename_column = FALSE)
}
\arguments{
\item{conn}{A DuckDB connection.}

\item{path}{Path(s) to parquet file(s).}

\item{filename_column}{A column name to the metadata that indicates which row came from which file.
By default it does not add the column.}
}
\value{
An SQL data frame
}
\description{
Returns a tibble from a parquet file path via DuckDB
Since dbplyr 2.4.0, raw file paths aren't handled very well
See: \url{https://github.com/duckdb/duckdb-r/issues/38}
Hence the need for this method
}
\keyword{internal}
