% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_cell_communication_strength}
\alias{get_cell_communication_strength}
\title{Retrieve cellNexus cell communication ligand–receptor strength as a data frame.}
\usage{
get_cell_communication_strength(
  cloud_metadata = get_metadata_url("cellNexus_lr_signaling_pathway_strength.parquet"),
  local_metadata = NULL,
  cache_directory = get_default_cache_dir(),
  use_cache = TRUE
)
}
\arguments{
\item{cloud_metadata}{Character vector of any length. HTTP URL/URLs pointing
to the name and location of parquet database/databases. By default, it points to
cell communication metadata in cellNexus ARDC Nectar Research Cloud.
Assign \code{NULL} to query local_metadata only if exists.}

\item{local_metadata}{Optional character vector of any length representing the local
path of parquet database(s).}

\item{cache_directory}{Optional character vector of length 1. A file path on
your local system to a directory (not a file) that will be used to store
\code{metadata.parquet}}

\item{use_cache}{Optional logical scalar. If \code{TRUE} (the default), and this
function has been called before with the same parameters, then a cached
reference to the table will be returned. If \code{FALSE}, a new connect138/4.7ion will
be created no matter what.}
}
\value{
A lazy data.frame subclass containing the metadata. You can interact
with this object using most standard dplyr functions. For string matching,
it is recommended that you use \code{stringr::str_like} to filter character
columns, as \code{stringr::str_match} will not work.
}
\description{
Downloads a parquet database of the cell communication strength to a local
cache, and then opens it as a data frame. It can then be filtered.
}
\details{
The returned table integrates three levels of cell communication
inference from \code{CellChat}, for each sample:
(i) ligand–receptor–level communication (\code{lr_prob}, \code{lr_pval}),
(ii) pathway-level aggregated signaling (\code{pathway_prob}, \code{pathway_pval}),
(iii) cell-pair–level summaries of communication breadth
(\code{interaction_count} - number of significant LR interactions) and
intensity (\code{interaction_weight} - overall communication strength).

Together, these metrics allow simultaneous assessment of signaling specificity,
pathway dominance, and global communication structure between cell populations.
}
\examples{
# For fast build purpose only, you do not need to specify anything in cloud_metadata.
communication_meta <- get_cell_communication_strength(cloud_metadata = SAMPLE_DATABASE_URL)
}
