% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_per_million.R
\name{get_counts_per_million}
\alias{get_counts_per_million}
\title{Generating counts per million from a SingleCellExperiment object}
\source{
\href{https://www.biorxiv.org/content/10.64898/2026.04.14.718336v3}{Shen et al.,2026}
}
\usage{
get_counts_per_million(sce, output_file)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{output_file}{A character vector of CPM Anndata file path}
}
\value{
A directory stores counts per million Anndata
}
\description{
Generating counts per million from a SingleCellExperiment object
}
\examples{
data(pbmc3k_sce)
get_counts_per_million(pbmc3k_sce, tempfile(fileext = ".h5ad"))
}
\references{
Shen, M., Y. Gao, N. Liu, D. Bhuva, M. Milton, J. Henao,
J. Andrews, E. Yang, C. Zhan, N. Liu, S. Si, J. W. Hutchison,
M. H. Shakeel, M. Morgan, A. T. Papenfuss, J. Iskander, J. M. Polo,
and S. Mangiola. "cellNexus: Quality control, annotation, aggregation
and analytical layers for the Human Cell Atlas data." bioRxiv (2026).
doi:10.64898/2026.04.14.718336.
}
