% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_data_container}
\alias{get_data_container}
\title{Gets data from curated metadata container}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
get_data_container(
  data,
  assays = "counts",
  atlas_name,
  cell_aggregation,
  cache_directory = get_default_cache_dir(),
  repository = COUNTS_URL,
  grouping_column,
  features = NULL
)
}
\arguments{
\item{data}{A data frame containing, at minimum, \code{cell_id}, \code{file_id_cellNexus_single_cell}
and/or \code{file_id_cellNexus_pseudobulk} column, which correspond to a single cell ID,
file subdivision for internal use for single_cell and/or pseudobulk level.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}

\item{assays}{A character vector specifying the desired assay(s) to be requested.
Valid elements include "counts", "cpm", and "rank" for single-cell analyses, or
"counts" and "quantile_normalised" for pseudobulk analyses.
The default setting retrieves only the counts assay.
If your analysis involves a smaller set of genes, consider using the "cpm" assay.
The "rank" assay is suited for signature calculations across millions of cells.}

\item{atlas_name}{A character string specifying the name of the atlas to query.}

\item{cell_aggregation}{A character vector that specifies which cell aggregation strategy
should be applied. This will create  a corresponding subdirectory in the atlas cache.
Choose one of the following: single_cell, pseudobulk}

\item{cache_directory}{An optional character vector of length one. If
provided, it should indicate a local file path where any remotely accessed
files should be copied.}

\item{repository}{A character vector of length one. If provided, it should be
an HTTP URL pointing to the location where the single cell data is stored.}

\item{grouping_column}{A character vector of metadata column for grouping. "file_id_cellNexus_single_cell" for
for Single Cell Experiment, "file_id_cellNexus_pseudobulk" for pseudobulk.}

\item{features}{An optional character vector of features (ie genes) to return
the counts for. By default counts for all features will be returned.}
}
\description{
Given a data frame of Curated Atlas metadata obtained from \code{\link[=get_metadata]{get_metadata()}},
returns a \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment}} object
corresponding to the samples in that data frame
}
\references{
Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen,
A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human
immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06.
doi:10.1101/2023.06.08.542671.
}
