% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{get_metacell_metadata_url}
\alias{get_metacell_metadata_url}
\title{Returns the URL for metacell metadata file}
\format{
An object of class \code{character} of length 1.
}
\usage{
get_metacell_metadata_url
}
\description{
Returns the URL for metacell metadata file
}
\keyword{datasets}
