% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{group_to_data_container}
\alias{group_to_data_container}
\title{Converts a data frame into a Summarized Experiment}
\usage{
group_to_data_container(
  i,
  df,
  dir_prefix,
  features,
  grouping_column,
  metacell_column = NULL
)
}
\arguments{
\item{i}{Suffix to be added to the column names, to make them unique}

\item{df}{The data frame to be converted}

\item{dir_prefix}{The path to the single cell experiment, minus the final
segment}

\item{features}{The list of genes/rows of interest}

\item{grouping_column}{A character vector of metadata column for grouping}

\item{metacell_column}{A character vector of metacell column (e.g. "metacell_2", "metacell_4") from metadata.}
}
\value{
A \code{SummarizedExperiment} object
}
\description{
This function converts a given data frame into a Summarized Experiment object,
allowing for handling of single-cell or pseudobulk data based on specified experiment
type. It requires specific columns for the data frame based on the type of
experiment data being processed.
}
\keyword{internal}
