% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{join_census_table}
\alias{join_census_table}
\title{Join Census metadata to an existing data frame}
\usage{
join_census_table(
  tbl,
  cloud_metadata = get_metadata_url("census_cell_metadata.2.1.0.parquet"),
  cache_directory = get_default_cache_dir(),
  join_keys = c("sample_id", "dataset_id", "observation_joinid")
)
}
\arguments{
\item{tbl}{A \code{tbl_sql} object (from get_metadata) or a database connection.}

\item{cloud_metadata}{HTTP URL/URLs pointing to the census parquet database name and location.}

\item{cache_directory}{A character string specifying the local cache
directory where remote parquet files will be stored. Defaults to
\code{\link[=get_default_cache_dir]{get_default_cache_dir()}}.}

\item{join_keys}{A character vector of column names used for the join.
Defaults to \code{c("sample_id", "dataset_id", "observation_joinid")}.}
}
\value{
A lazy SQL table with Census metadata joined to the cellNexus metadata.
}
\description{
Downloads and joins the Census metadata with cellNexus metadata.
This function creates indexed tables for efficient joining and returns a data frame.
}
\examples{
library(dplyr)
get_metadata(cloud_metadata = SAMPLE_DATABASE_URL) |> head(2) |>
  # You do not need to specify anything in cloud_metadata
  join_census_table(
    cloud_metadata = SAMPLE_DATABASE_URL,
    cache_directory = tempdir()
  )
}
