% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metacell_counts}
\alias{metacell_counts}
\title{Sample Metacell SingleCellExperiment Object}
\format{
An object of class \code{SingleCellExperiment} with:
\describe{
\item{assays}{Gene expression matrix with counts assay aggregated into metacells}
\item{colData}{Metacell metadata including metacell_2, etc.}
}
}
\source{
Created from cellNexus datasets
}
\usage{
metacell_counts
}
\description{
A pre-made SingleCellExperiment object with metacell aggregated data for vignette demonstration.
This object is used in the vignette to avoid downloading data during package build.
}
\details{
See \code{dev/create_vignette_data.R} for the creation script.
}
\examples{
data(metacell_counts)
metacell_counts
}
\keyword{datasets}
