% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc3k_sce}
\alias{pbmc3k_sce}
\title{Sample SingleCellExperiment Object}
\format{
An object of class \code{SingleCellExperiment} with:
\describe{
\item{assays}{Gene expression matrix with Ensembl gene IDs as rownames}
\item{colData}{Cell metadata including sample_id, cell_type_unified_ensemble, nCount_RNA, etc.}
\item{metadata}{List containing 'data' field with cellNexus-formatted metadata including:
\itemize{
\item cell_id: Unique cell identifier
\item sample_id: Sample identifier
\item cell_type_unified_ensemble: Cell type annotation
\item nCount_RNA: Number of RNA molecules per cell
\item ident: Seurat cluster identity
\item dataset_id: Dataset identifier
\item file_id_cellNexus_single_cell: Generated file ID for cellNexus
\item atlas_id: Atlas identifier with date
}
}
}
}
\source{
Created from pbmc3k dataset (SeuratData package)

\href{https://www.biorxiv.org/content/10.64898/2026.04.14.718336v3}{Shen et al.,2026}
}
\description{
A sample SingleCellExperiment object created from the pbmc3k dataset for testing and demonstration purposes.
The dataset contains 500 cells with gene expression data mapped to Ensembl gene IDs and formatted
with cellNexus-compatible metadata structure.
}
\details{
See \code{dev/create_pbmc3k_sce.R} for the complete creation script.
}
\examples{
data(pbmc3k_sce)
pbmc3k_sce
# Access metadata
S4Vectors::metadata(pbmc3k_sce)$data
}
\references{
Shen, M., Y. Gao, N. Liu, D. Bhuva, M. Milton, J. Henao,
J. Andrews, E. Yang, C. Zhan, N. Liu, S. Si, J. W. Hutchison,
M. H. Shakeel, M. Morgan, A. T. Papenfuss, J. Iskander, J. M. Polo,
and S. Mangiola. "cellNexus: Quality control, annotation, aggregation
and analytical layers for the Human Cell Atlas data." bioRxiv (2026).
doi:10.64898/2026.04.14.718336.
}
\keyword{datasets}
