% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudobulk_counts}
\alias{pseudobulk_counts}
\title{Sample Pseudobulk SingleCellExperiment Object}
\format{
An object of class \code{SingleCellExperiment} with:
\describe{
\item{assays}{Gene expression matrix with counts assay aggregated by sample and cell type}
\item{colData}{Sample metadata including sample_id, cell_type_unified_ensemble, etc.}
}
}
\source{
Created from cellNexus datasets
}
\usage{
pseudobulk_counts
}
\description{
A pre-made SingleCellExperiment object with pseudobulk aggregated data for vignette demonstration.
This object is used in the vignette to avoid downloading data during package build.
}
\details{
See \code{dev/create_vignette_data.R} for the creation script.
}
\examples{
data(pseudobulk_counts)
pseudobulk_counts
}
\keyword{datasets}
