% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_and_join_parquets}
\alias{read_and_join_parquets}
\title{Join multiple parquet files with indexing}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
read_and_join_parquets(conn, path, join_keys, filename_column = FALSE)
}
\arguments{
\item{conn}{Database connection}

\item{path}{Vector of parquet file paths}

\item{join_keys}{Columns to join on}

\item{filename_column}{Whether to include filename column}
}
\value{
Lazy SQL table
}
\description{
Join multiple parquet files with indexing
}
\keyword{internal}
