% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_sce_as_h5ad}
\alias{save_sce_as_h5ad}
\title{Save a SingleCellExperiment as an AnnData file}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
save_sce_as_h5ad(sce, path, ...)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{path}{Output file path for the \code{.h5ad} file.}

\item{...}{
  Arguments passed on to \code{\link[zellkonverter:writeH5AD]{zellkonverter::writeH5AD}}
  \describe{
    \item{\code{file}}{String containing a path to write the new \code{.h5ad} file.}
    \item{\code{X_name}}{Name of the assay to use as the primary matrix (\code{X}) of the
AnnData object. If \code{NULL}, the first assay of \code{sce} will be used by default.}
    \item{\code{skip_assays}}{Logical scalar indicating whether assay matrices should
be ignored when writing to \code{file}.}
    \item{\code{compression}}{Type of compression when writing the new \code{.h5ad} file.}
    \item{\code{version}}{A string giving the version of the \strong{anndata} Python library
to use. Allowed values are available in \code{.AnnDataVersions}. By default the
latest version is used.}
    \item{\code{verbose}}{Logical scalar indicating whether to print progress messages.
If \code{NULL} uses \code{getOption("zellkonverter.verbose")}.}
  }}
}
\value{
Called for its side effect of writing an \code{.h5ad} file.
}
\description{
Reports and removes \code{colData} columns that are entirely \code{NA} before writing to \code{.h5ad}.
If the object contains only one column, the assay is duplicated to avoid
single-column export issues.
}
\examples{
data("pbmc3k_sce")
save_sce_as_h5ad(pbmc3k_sce, tempfile(fileext=".h5ad"))
}
