% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_counts}
\alias{single_cell_counts}
\title{Sample SingleCellExperiment Object with Counts Assay}
\format{
An object of class \code{SingleCellExperiment} with:
\describe{
\item{assays}{Gene expression matrix with counts assay}
\item{colData}{Cell metadata including sample_id, cell_type_unified_ensemble, etc.}
}
}
\source{
Created from cellNexus datasets
}
\usage{
single_cell_counts
}
\description{
A pre-made SingleCellExperiment object with counts assay for vignette demonstration.
This object is used in the vignette to avoid downloading data during package build.
}
\details{
See \code{dev/create_vignette_data.R} for the creation script.
}
\keyword{datasets}
