% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_cell_cpm}
\alias{single_cell_cpm}
\title{Sample SingleCellExperiment Object with CPM Assay}
\format{
An object of class \code{SingleCellExperiment} with:
\describe{
\item{assays}{Gene expression matrix with cpm assay}
\item{colData}{Cell metadata including sample_id, cell_type_unified_ensemble, etc.}
}
}
\source{
Created from cellNexus datasets
}
\description{
A pre-made SingleCellExperiment object with counts-per-million (CPM) assay for vignette demonstration.
This object is used in the vignette to avoid downloading data during package build.
}
\details{
See \code{dev/create_vignette_data.R} for the creation script.
}
\examples{
data(single_cell_cpm)
single_cell_cpm
}
\keyword{datasets}
