% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{single_line_str}
\alias{single_line_str}
\title{Formats a multi-line string as it it were on one line}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
single_line_str(text)
}
\arguments{
\item{text}{Any character vector}
}
\value{
The same character vector, with newlines and subsequent whitespace
removed
}
\description{
Formats a multi-line string as it it were on one line
}
\keyword{internal}
