% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{sync_assay_files}
\alias{sync_assay_files}
\title{Synchronises one or more remote assays with a local copy}
\usage{
sync_assay_files(
  url = parse_url(COUNTS_URL),
  atlas_name,
  cell_aggregation,
  cache_dir,
  subdirs,
  files
)
}
\arguments{
\item{url}{A character vector of length one. The base HTTP URL from which to
obtain the files.}

\item{cache_dir}{A character vector of length one. The local filepath to
synchronise files to.}

\item{subdirs}{A character vector of subdirectories within the root URL to
sync. These correspond to assays.}

\item{files}{A character vector containing one or more file_id_cellNexus_single_cell entries}
}
\value{
A character vector consisting of file paths to all the newly
downloaded files

A character vector of files that have been downloaded
}
\description{
Synchronises one or more remote assays with a local copy
}
\keyword{internal}
