% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{upload_swift}
\alias{upload_swift}
\title{Upload a file to the Nectar object store}
\usage{
upload_swift(
  source,
  container,
  name = basename(source),
  credential_id = NULL,
  credential_secret = NULL
)
}
\arguments{
\item{source}{A character scalar indicating the local path to the file to
upload}

\item{container}{A character scalar indicating the name of the container to
upload to}

\item{name}{An optional character scalar indicating the name the file should
have after being uploaded. Defaults to being the basename of the source
file.}

\item{credential_id}{The OpenStack application credential secret as a
character scalar}
}
\value{
\code{NULL}, invisibly
}
\description{
Upload a file to the Nectar object store
}
\examples{
\dontrun{
upload_swift(
  "/vast/projects/cellxgene_curated/metadata_parquet_0.2",
  "cellNexus-metadata",
  credential_id = "ABCDEFGHIJK",
  credential_secret = "ABCD1234EFGH-5678IJK"
)
}
}
\keyword{internal}
