% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_file_size}
\alias{url_file_size}
\title{Gets the file size of a number of remote files}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
url_file_size(urls)
}
\arguments{
\item{urls}{A character vector containing URLs}
}
\value{
The file size of each of the files pointed to by the provided URL,
in gigabytes, as double vector
}
\description{
Gets the file size of a number of remote files
}
\keyword{internal}
