% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_index_parquet}
\alias{add_index_parquet}
\title{Create join index from a parquet file path}
\usage{
add_index_parquet(
  conn,
  path,
  index_cols,
  table_name = NULL,
  filename_column = FALSE
)
}
\description{
Create join index from a parquet file path
}
