% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_index_parquet}
\alias{add_index_parquet}
\title{Create indexed table from parquet file}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
add_index_parquet(
  conn,
  path,
  index_cols,
  table_name = NULL,
  filename_column = FALSE
)
}
\arguments{
\item{conn}{Database connection}

\item{path}{Path to parquet file}

\item{index_cols}{Columns to create index on}

\item{table_name}{Name for the table (default: filename without extension)}

\item{filename_column}{Whether to include filename column}
}
\value{
Table name (invisibly)
}
\description{
Create indexed table from parquet file
}
\keyword{internal}
