% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_metadata_and_counts.R
\name{calculate_pseudobulk}
\alias{calculate_pseudobulk}
\title{Generate pseudobulk counts and quantile_normalised counts}
\usage{
calculate_pseudobulk(
  sce_data,
  atlas_name,
  import_date,
  cell_aggregation,
  cache_dir = get_default_cache_dir()
)
}
\arguments{
\item{sce_data}{A SingleCellExperiment object, the metadata slot of which
must contain \code{cell_id} and \code{dataset_id}}

\item{atlas_name}{A character string specifying the name of the atlas to import.}

\item{import_date}{A character vector that specifies the date of the import.
The date should be in the international format 'YYYY-MM-DD' for clarity and consistency.}

\item{cell_aggregation}{A character vector that specifies which cell aggregation strategy
should be applied. This will create  a corresponding subdirectory in the atlas cache.
Choose one of the following: single_cell, pseudobulk.}

\item{cache_dir}{Optional character vector of length 1. A file path on
your local system to a directory (not a file) that will be used to store pseudobulk counts}
}
\value{
Pseudobulk counts in \code{Anndata} format stored in the cache directory
}
\description{
Generate pseudobulk counts and quantile_normalised counts
}
\examples{
data(pbmc3k_sce)
calculate_pseudobulk(pbmc3k_sce,
                   atlas_name = "pbmc3k_sce_atlas",
                   import_date = format(Sys.Date(), "\%d-\%m-\%Y"),
                   cell_aggregation = "pseudobulk",
                   cache_dir = get_default_cache_dir())
}
