% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{cellNexus-package}
\alias{cellNexus-package}
\alias{cellNexus}
\title{cellNexus: Query Interface for the Human Cell Atlas}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\value{
The cellNexus package (invisibly).
}
\description{
\code{cellNexus} provides a query interface for programmatic exploration
and retrieval of the harmonised, curated and reannotated CELLxGENE
single-cell Human Cell Atlas. The package allows users to query metadata
and download single-cell RNA sequencing data in various formats including
SingleCellExperiment, Seurat, and pseudobulk SummarizedExperiment objects.
}
\section{Getting Started}{

To get started with \code{cellNexus}, first load the package and retrieve
the metadata:

\preformatted{
library(cellNexus)
metadata <- get_metadata()
}

Then filter the metadata to find cells of interest and download the data:

\preformatted{
filtered_metadata <- metadata |>
    dplyr::filter(
        imputed_ethnicity == "African" &
        cell_type_unified_ensemble \%LIKE\% "\%cd4\%"
    )

sce <- get_single_cell_experiment(filtered_metadata)
}
}

\section{Main Functions}{

\describe{
\item{\code{\link{get_metadata}}}{Retrieve and query the harmonised metadata}
\item{\code{\link{get_single_cell_experiment}}}{Download data as SingleCellExperiment objects}
\item{\code{\link{get_seurat}}}{Download data as Seurat objects}
\item{\code{\link{get_pseudobulk}}}{Download aggregated pseudobulk data}
}
}

\section{Data Licensing}{

\strong{Important:} The Human Cell Atlas data accessed through this package
is subject to its own licensing terms, which differ from the package license.
The \code{cellNexus} package itself is licensed under GPL-3. However, the
underlying Human Cell Atlas data is typically licensed under Creative Commons
Attribution (CC-BY) or similar open data licenses as specified by the
Human Cell Atlas Data Use Agreement. Users should review the specific
licensing terms for any datasets they access through this package.
For more information, see the Human Cell Atlas Data Portal:
\url{https://data.humancellatlas.org/about}
}

\section{Vignettes}{

See \code{vignette("cellNexus", package = "cellNexus")} for a comprehensive
introduction to using the package.
}

\references{
Shen, M., Y. Gao, N. Liu, D. Bhuva, M. Milton, J. Henao,
J. Andrews, E. Yang, C. Zhan, N. Liu, S. Si, J. W. Hutchison,
M. H. Shakeel, M. Morgan, A. T. Papenfuss, J. Iskander, J. M. Polo,
and S. Mangiola. "cellNexus: Quality control, annotation, aggregation
and analytical layers for the Human Cell Atlas data." bioRxiv (2026).
doi:10.64898/2026.04.14.718336.
}
