% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_app.R
\name{create_interface_app}
\alias{create_interface_app}
\title{Create a Shiny app that allows users to generate filtering & retreival code for cellNexus}
\usage{
create_interface_app(ui_choices)
}
\arguments{
\item{ui_choices}{A list of pre-computed unique values for each filterable column in the metadata.}
}
\value{
A Shiny app that allows users to filter cellNexus metadata and generate code for retrieval
in the selected format.
}
\description{
Create a Shiny app that allows users to generate filtering & retreival code for cellNexus
}
\examples{
\dontrun{
library(cellNexus)
app <- create_interface_app(ui_choices)
shiny::runApp(app)
}
}
\author{
Jared Andrews
}
