% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_app.R
\name{create_interface_app}
\alias{create_interface_app}
\title{Create a Shiny app that allows users to generate filtering & retreival code for cellNexus}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
create_interface_app(metadata)
}
\arguments{
\item{metadata}{cellNexus metadata as returned by \code{get_metadata()}.}
}
\value{
A Shiny app that allows users to filter cellNexus metadata and generate code for retrieval
in the selected format.
}
\description{
Create a Shiny app that allows users to generate filtering & retreival code for cellNexus
}
\examples{
get_default_cache_dir()
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Create the interface app with metadata
metadata <- get_metadata(cloud_metadata = SAMPLE_DATABASE_URL)
app <- create_interface_app(metadata)
# Run the app
shiny::runApp(app)
\dontshow{\}) # examplesIf}
}
\author{
Jared Andrews
}
