% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_app.R
\name{create_interface_app}
\alias{create_interface_app}
\title{Create a Shiny app that allows users to generate filtering & retreival code for cellNexus}
\usage{
create_interface_app(metadata)
}
\arguments{
\item{metadata}{cellNexus metadata as returned by \code{get_metadata()}.}
}
\value{
A Shiny app that allows users to filter cellNexus metadata and generate code for retrieval
in the selected format.
}
\description{
Create a Shiny app that allows users to generate filtering & retreival code for cellNexus
}
\author{
Jared Andrews
}
