% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_per_million.R
\name{get_counts_per_million}
\alias{get_counts_per_million}
\title{Generating counts per million from a SingleCellExperiment object}
\usage{
get_counts_per_million(sce, output_file)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{output_file}{A character vector of CPM Anndata file path}
}
\value{
A directory stores counts per million Anndata
}
\description{
Generating counts per million from a SingleCellExperiment object
}
\examples{
data(pbmc3k_sce)
get_counts_per_million(pbmc3k_sce, tempfile(fileext = ".h5ad"))
}
