% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_per_million.R
\name{get_counts_per_million}
\alias{get_counts_per_million}
\title{Generating counts per million from a SingleCellExperiment object}
\usage{
get_counts_per_million(sce, input_file, output_file)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{input_file}{A character vector of counts Anndata file path}

\item{output_file}{A character vector of CPM Anndata file path}
}
\value{
A directory stores counts per million Anndata
}
\description{
Generating counts per million from a SingleCellExperiment object
}
\keyword{internal}
