% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default_cache_dir}
\alias{get_default_cache_dir}
\title{Returns the default cache directory with a version number}
\source{
\href{https://www.biorxiv.org/content/10.64898/2026.04.14.718336v3}{Shen et al.,2026}
}
\usage{
get_default_cache_dir()
}
\value{
A length one character vector.
}
\description{
Returns the default cache directory with a version number
}
\examples{
get_metadata(cloud_metadata = SAMPLE_DATABASE_URL, cache_directory = get_default_cache_dir())
}
\references{
Shen, M., Y. Gao, N. Liu, D. Bhuva, M. Milton, J. Henao,
J. Andrews, E. Yang, C. Zhan, N. Liu, S. Si, J. W. Hutchison,
M. H. Shakeel, M. Morgan, A. T. Papenfuss, J. Iskander, J. M. Polo,
and S. Mangiola. "cellNexus: Quality control, annotation, aggregation
and analytical layers for the Human Cell Atlas data." bioRxiv (2026).
doi:10.64898/2026.04.14.718336.
}
