% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_metacell}
\alias{get_metacell}
\title{Gets a Metacell from curated metadata}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
get_metacell(
  data,
  assays = "counts",
  cell_aggregation,
  cache_directory = get_default_cache_dir(),
  repository = COUNTS_URL,
  features = NULL
)
}
\arguments{
\item{data}{A data frame containing, at minimum, \code{sample_id}, \code{file_id_cellNexus_single_cell},
\code{atlas_id} and a metacell column (e.g \code{metacell_2}) columns, which correspond to sample ID
file subdivision for internal use, atlas name in format (e.g cellxgene/06-02-2025)
for internal use, and metacell column to be queried.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.}

\item{assays}{A character vector of metacell counts. Default to "counts".}

\item{cell_aggregation}{A character vector representing the level of metacell aggregation.
It indicates a group of cells that can be divided into the number of metacells.
Each metacell comprises a minimum of ten single cells by default.}

\item{cache_directory}{An optional character vector of length one. If
provided, it should indicate a local file path where any remotely accessed
files should be copied.}

\item{repository}{A character vector of length one. If provided, it should be
an HTTP URL pointing to the location where the single cell data is stored.}

\item{features}{An optional character vector of features (ie genes) to return
the counts for. By default counts for all features will be returned.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Given a data frame of Curated Atlas metadata obtained from \code{\link[=get_metadata]{get_metadata()}},
returns a \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}} object
corresponding to the samples in that data frame
}
\examples{
meta <- get_metadata() |> head(2)
metacell <- meta |> get_metacell(cell_aggregation = "metacell_2")
}
\references{
Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen,
A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human
immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06.
doi:10.1101/2023.06.08.542671.
}
