% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_url}
\alias{get_metadata_url}
\title{Returns the URLs for all metadata files}
\source{
\href{https://www.biorxiv.org/content/10.64898/2026.04.14.718336v3}{Shen et al.,2026}
}
\usage{
get_metadata_url(
  databases = c("hca2024_v2.3.0.parquet", "hca2025_v1.0.0.parquet")
)
}
\arguments{
\item{databases}{A character vector specifying the names of the metadata files.
Download the specific metadata by defining the metadata version.}
}
\value{
A character vector of URLs to parquet files to download
}
\description{
Returns the URLs for all metadata files
}
\examples{
get_metadata_url("hca2024_v2.3.0.parquet")
}
\references{
Shen, M., Y. Gao, N. Liu, D. Bhuva, M. Milton, J. Henao,
J. Andrews, E. Yang, C. Zhan, N. Liu, S. Si, J. W. Hutchison,
M. H. Shakeel, M. Morgan, A. T. Papenfuss, J. Iskander, J. M. Polo,
and S. Mangiola. "cellNexus: Quality control, annotation, aggregation
and analytical layers for the Human Cell Atlas data." bioRxiv (2026).
doi:10.64898/2026.04.14.718336.
}
