% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_metadata_url}
\alias{get_metadata_url}
\title{Returns the URLs for all metadata files}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
get_metadata_url(
  databases = "metadata.1.2.13.parquet",
  use_split_files = FALSE,
  use_census = FALSE,
  use_metacell = FALSE
)
}
\arguments{
\item{databases}{A character vector specifying the names of the metadata files.
Download the specific metadata by defining the metadata version. By default, it uses
metadata.1.2.13.parquet}
}
\value{
A character vector of URLs to parquet files to download
}
\description{
Returns the URLs for all metadata files
}
\examples{
get_metadata_url("metadata.1.2.13.parquet")
}
\references{
Mangiola, S., M. Milton, N. Ranathunga, C. S. N. Li-Wai-Suen,
A. Odainic, E. Yang, W. Hutchison et al. "A multi-organ map of the human
immune system across age, sex and ethnicity." bioRxiv (2023): 2023-06.
doi:10.1101/2023.06.08.542671.
}
