% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{get_SingleCellExperiment}
\alias{get_SingleCellExperiment}
\title{Gets a SingleCellExperiment from curated metadata}
\source{
\href{https://www.biorxiv.org/content/10.64898/2026.04.14.718336v3}{Shen et al.,2026}
}
\usage{
get_SingleCellExperiment(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_single_cell_experiment]{get_single_cell_experiment}}
  \describe{
    \item{\code{data}}{A data frame containing, at minimum, \code{cell_id}, \code{file_id_cellNexus_single_cell}
and \code{atlas_id} columns, which correspond to a single cell ID, file subdivision for internal use,
and atlas name in format (e.g cellxgene_2024/0.1.0) for internal use.
They can be obtained from the \code{\link[=get_metadata]{get_metadata()}} function.
Use \code{get_atlas_versions()} to download atlas versions data frame.}
    \item{\code{assays}}{A character vector specifying the desired assay(s) to be requested.
Valid elements include "counts", "cpm", "rank", and "sct" for single-cell analyses
The default setting retrieves only the counts assay.
If your analysis involves a smaller set of genes, consider using the "cpm" assay.
The "rank" assay is suited for signature calculations across millions of cells.}
    \item{\code{cell_aggregation}}{A character vector that specifies which cell aggregation
strategy should be applied. This will create a corresponding subdirectory
in the cache directory. Single cell level is applied by default.}
    \item{\code{cache_directory}}{An optional character vector of length one. If
provided, it should indicate a local file path where any remotely accessed
files should be copied.}
    \item{\code{repository}}{A character vector of length one. If provided, it should be
an HTTP URL pointing to the location where the single cell data is stored.}
    \item{\code{features}}{An optional character vector of features (ie genes) to return
the counts for. By default counts for all features will be returned.}
    \item{\code{download_only}}{Logical scalar. When \code{TRUE}, remote files are
synchronised to \code{cache_directory} but the data is not read or assembled
into an in-memory object. Returns \code{invisible(NULL)}. Useful for
pre-fetching datasets only.}
  }}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
Given a data frame of Curated Atlas metadata obtained from \code{\link[=get_metadata]{get_metadata()}},
returns a \code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}} object
corresponding to the samples in that data frame
}
\examples{
# Use the lightweight sample database URL (for fast checks during development only)
meta <- get_metadata(cloud_metadata = cellNexus::SAMPLE_DATABASE_URL) |> head(2)
sce <- get_single_cell_experiment(meta)
}
\references{
Shen, M., Y. Gao, N. Liu, D. Bhuva, M. Milton, J. Henao,
J. Andrews, E. Yang, C. Zhan, N. Liu, S. Si, J. W. Hutchison,
M. H. Shakeel, M. Morgan, A. T. Papenfuss, J. Iskander, J. M. Polo,
and S. Mangiola. "cellNexus: Quality control, annotation, aggregation
and analytical layers for the Human Cell Atlas data." bioRxiv (2026).
doi:10.64898/2026.04.14.718336.
}
