% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_quality_cells}
\alias{keep_quality_cells}
\title{Keep high-quality cells based on QC columns}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
keep_quality_cells(
  data,
  empty_droplet_col = "empty_droplet",
  alive_col = "alive",
  doublet_col = "scDblFinder.class"
)
}
\arguments{
\item{data}{A data frame or tibble containing single-cell metadata.}

\item{empty_droplet_col}{A string specifying the column name
that indicates empty droplets (default: \code{"empty_droplet"}).
Expected logical vector}

\item{alive_col}{A string specifying the column name
that indicates whether cells are alive (default: \code{"alive"}).
Expected logical vector}

\item{doublet_col}{A string specifying the column name
that indicates doublets (default: \code{"scDblFinder.class"}).
Expected character vector: \code{"doublet"} and/or \code{"singlet"} and/or \code{"unknown"}.}
}
\value{
A filtered data frame containing only cells that pass all QC checks.
}
\description{
Keep high-quality cells based on QC columns
}
\examples{
get_metadata(cloud_metadata = SAMPLE_DATABASE_URL, cache_directory = tempdir()) |> 
  head(2) |>
  keep_quality_cells()
}
