% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_and_join_parquets}
\alias{read_and_join_parquets}
\title{Join registered databased based on index}
\usage{
read_and_join_parquets(conn, path, join_keys, filename_column = FALSE)
}
\description{
Join registered databased based on index
}
