% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{SAMPLE_DATABASE_URL}
\alias{SAMPLE_DATABASE_URL}
\title{URL pointing to the sample metadata file, which is smaller and for test,
demonstration, and vignette purposes only}
\format{
An object of class \code{character} of length 1.
}
\source{
\href{https://www.biorxiv.org/content/10.64898/2026.04.14.718336v3}{Shen et al.,2026}
}
\usage{
SAMPLE_DATABASE_URL
}
\value{
Character scalar consisting of the URL/URLs
}
\description{
URL pointing to the sample metadata file, which is smaller and for test,
demonstration, and vignette purposes only
}
\examples{
get_metadata(cloud_metadata = SAMPLE_DATABASE_URL, cache_directory = tempdir())
}
\references{
Shen, M., Y. Gao, N. Liu, D. Bhuva, M. Milton, J. Henao,
J. Andrews, E. Yang, C. Zhan, N. Liu, S. Si, J. W. Hutchison,
M. H. Shakeel, M. Morgan, A. T. Papenfuss, J. Iskander, J. M. Polo,
and S. Mangiola. "cellNexus: Quality control, annotation, aggregation
and analytical layers for the Human Cell Atlas data." bioRxiv (2026).
doi:10.64898/2026.04.14.718336.
}
\keyword{datasets}
