% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_sce_as_h5ad}
\alias{save_sce_as_h5ad}
\title{Save a SingleCellExperiment as an AnnData file}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
save_sce_as_h5ad(sce, path, ...)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{path}{Output file path for the \code{.h5ad} file.}

\item{...}{
  Arguments passed on to \code{\link[anndataR:write_h5ad]{anndataR::write_h5ad}}
  \describe{
    \item{\code{object}}{The object to write, either a
\code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment::SingleCellExperiment}} or a
\code{\link[SeuratObject:Seurat-class]{SeuratObject::Seurat}} object}
    \item{\code{compression}}{The compression algorithm to use when writing the HDF5
file. Can be one of \code{"none"}, \code{"gzip"} or \code{"lzf"}. Defaults to \code{"none"}.}
    \item{\code{mode}}{The mode to open the HDF5 file.
\itemize{
\item \code{a} creates a new file or opens an existing one for read/write
\item \verb{r+} opens an existing file for read/write
\item \code{w} creates a file, truncating any existing ones
\item \verb{w-}/\code{x} are synonyms creating a file and failing if it already exists
}}
  }}
}
\value{
Called for its side effect of writing an \code{.h5ad} file.
}
\description{
Reports and removes \code{colData} columns that are entirely \code{NA} before writing to \code{.h5ad}.
If the object contains only one column, the assay is duplicated to avoid
single-column export issues.
}
\examples{
data("pbmc3k_sce")
save_sce_as_h5ad(pbmc3k_sce, tempfile(fileext=".h5ad"))
}
