% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_sce_as_h5ad}
\alias{save_sce_as_h5ad}
\title{Save SingleCellExperiment as AnnData (.h5ad) after cleaning NA columns from colData}
\source{
\href{https://www.biorxiv.org/content/10.1101/2023.06.08.542671v3}{Mangiola et al.,2023}
}
\usage{
save_sce_as_h5ad(sce, path)
}
\value{
\code{NULL}, invisible
}
\description{
Save SingleCellExperiment as AnnData (.h5ad) after cleaning NA columns from colData
}
\keyword{internal}
